Rename reader to file input
