+++ plugins/inputs/file/dev/docker-compose.yml
