+++ plugins/inputs/file/dev/json_a.log
