+++ plugins/inputs/file/dev/telegraf.conf
