+++ plugins/inputs/file/testfiles/grok_a.log
