+++ plugins/inputs/file/testfiles/json_a.log
